refactor: move bbox related structs into `summary.go`
